feat(cqrs): moved command bus to cqrs
