+++ crates/cqrs/src/bus/executor.rs
@@ -1,0 +1,1 @@
+pub struct Executor {}
