+++ crates/cqrs/src/bus/message.rs
@@ -1,0 +1,3 @@
+pub struct Message<T> {
+    payload: T,
+}
