+++ crates/cqrs/src/bus/mod.rs
@@ -1,0 +1,6 @@
+pub mod executor;
+pub mod message;
+
+pub struct Bus {
+    bound: usize,
+}
