--- crates/cqrs/src/error.rs
@@ -1,8 +1,0 @@
-use std::error::Error as StdError;
-use thiserror::Error;
-
-#[derive(Debug, Error)]
-pub enum AggregateError<T: StdError> {
-    #[error("{0}")]
-    UserError(T),
-}
